pyproject.toml
[project]
name = "fortissimusbellator"
version = "1.0.0"
description = "Fortissimus Bellator website"
authors = [
    {name = "Daniel Pinto",email = "dmp593@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "pillow (>=11.1.0,<12.0.0)",
    "django (>=5.1.7,<6.0.0)",
    "django-modeltranslation (>=0.19.12,<0.20.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "django-recaptcha (>=4.0.0,<5.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "django-colorfield (>=0.12.0,<0.13.0)",
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
